Make KtTestCompiler internal
