+++ detekt-generator/src/test/kotlin/io/gitlab/arturbosch/detekt/generator/util/CollectorTestExtensions.kt
@@ -1,10 +1,10 @@
 package io.gitlab.arturbosch.detekt.generator.util
 
+import io.github.detekt.test.utils.compileContentForTest
 import io.gitlab.arturbosch.detekt.generator.collection.Collector
-import io.github.detekt.test.utils.KtTestCompiler
 
 fun <T> Collector<T>.run(code: String): List<T> {
-    val ktFile = KtTestCompiler.compileFromContent(code.trimIndent())
+    val ktFile = compileContentForTest(code.trimIndent())
     visit(ktFile)
     return items
 }
