+++ detekt-test-utils/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/KotlinEnvironmentTestSetup.kt
@@ -1,12 +1,12 @@
 package io.gitlab.arturbosch.detekt.rules
 
-import io.github.detekt.test.utils.KtTestCompiler
+import io.github.detekt.test.utils.createEnvironment
 import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
 import org.spekframework.spek2.dsl.Root
 import org.spekframework.spek2.lifecycle.CachingMode
 
 fun Root.setupKotlinEnvironment() {
-    val wrapper = KtTestCompiler.createEnvironment()
+    val wrapper = createEnvironment()
 
     @Suppress("UNUSED_VARIABLE") // name is used for delegation
     val env: KotlinCoreEnvironment by memoized(CachingMode.EACH_GROUP) { wrapper.env }
