+++ detekt-test/src/main/kotlin/io/gitlab/arturbosch/detekt/test/RuleExtensions.kt
@@ -1,14 +1,18 @@
 package io.gitlab.arturbosch.detekt.test
 
 import io.github.detekt.test.utils.KotlinScriptEngine
-import io.github.detekt.test.utils.KtTestCompiler
+import io.github.detekt.test.utils.compileContentForTest
+import io.github.detekt.test.utils.compileForTest
 import io.gitlab.arturbosch.detekt.api.Finding
 import io.gitlab.arturbosch.detekt.api.Rule
 import io.gitlab.arturbosch.detekt.api.internal.BaseRule
 import org.intellij.lang.annotations.Language
 import org.jetbrains.kotlin.cli.jvm.compiler.KotlinCoreEnvironment
+import org.jetbrains.kotlin.cli.jvm.compiler.NoScopeRecordCliBindingTrace
+import org.jetbrains.kotlin.cli.jvm.compiler.TopDownAnalyzerFacadeForJVM
 import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.resolve.BindingContext
+import org.jetbrains.kotlin.resolve.lazy.declarations.FileBasedDeclarationProviderFactory
 import java.nio.file.Path
 
 private val shouldCompileTestSnippets: Boolean =
@@ -22,12 +26,12 @@ fun BaseRule.compileAndLint(@Language("kotlin") content: String): List<Finding> 
 }
 
 fun BaseRule.lint(@Language("kotlin") content: String): List<Finding> {
-    val ktFile = KtTestCompiler.compileFromContent(content.trimIndent())
+    val ktFile = compileContentForTest(content.trimIndent())
     return findingsAfterVisit(ktFile)
 }
 
 fun BaseRule.lint(path: Path): List<Finding> {
-    val ktFile = KtTestCompiler.compile(path)
+    val ktFile = compileForTest(path)
     return findingsAfterVisit(ktFile)
 }
 
@@ -38,12 +42,18 @@ fun BaseRule.compileAndLintWithContext(
     if (shouldCompileTestSnippets) {
         KotlinScriptEngine.compile(content)
     }
-    val ktFile = KtTestCompiler.compileFromContent(content.trimIndent())
-    val bindingContext = KtTestCompiler.getContextForPaths(environment, listOf(ktFile))
+    val ktFile = compileContentForTest(content.trimIndent())
+    val bindingContext = getContextForPaths(environment, listOf(ktFile))
     return findingsAfterVisit(ktFile, bindingContext)
 }
 
-fun BaseRule.lint(ktFile: KtFile) = findingsAfterVisit(ktFile)
+private fun getContextForPaths(environment: KotlinCoreEnvironment, paths: List<KtFile>) =
+    TopDownAnalyzerFacadeForJVM.analyzeFilesWithJavaIntegration(
+        environment.project, paths, NoScopeRecordCliBindingTrace(),
+        environment.configuration, environment::createPackagePartProvider, ::FileBasedDeclarationProviderFactory
+    ).bindingContext
+
+fun BaseRule.lint(ktFile: KtFile): List<Finding> = findingsAfterVisit(ktFile)
 
 private fun BaseRule.findingsAfterVisit(
     ktFile: KtFile,
@@ -54,12 +64,12 @@ private fun BaseRule.findingsAfterVisit(
 }
 
 fun Rule.format(@Language("kotlin") content: String): String {
-    val ktFile = KtTestCompiler.compileFromContent(content.trimIndent())
+    val ktFile = compileContentForTest(content.trimIndent())
     return contentAfterVisit(ktFile)
 }
 
 fun Rule.format(path: Path): String {
-    val ktFile = KtTestCompiler.compile(path)
+    val ktFile = compileForTest(path)
     return contentAfterVisit(ktFile)
 }
